+++ R/get-page-data.R
@@ -24,6 +24,7 @@
 #' @param title A character vector of page titles.
 #' @param language A character vector of two-letter language codes, either of
 #'   length 1 or the same length as `title`
+#' @param failure_mode Either "quiet" or "error." See [get_rest_resource()]
 #'
 #' @return A list, vector or tibble, the same length as `title`, with the
 #'   desired data.
@@ -32,36 +33,24 @@
 #'
 #' @examples
 #' # Get language links for a known page on English Wikipedia
-#' get_page_langlinks("Charles Harpur")
-#'
-#' # Many of these functions return a list of data frames. Tidyr can be useful.
-#' # Get 20 related pages for German City
-#' cities <- tibble::tribble(
-#'   ~city,
-#'   "Berlin",
-#'   "Darmstadt",
-#' ) %>%
-#'   dplyr::mutate(related = get_page_related(city))
-#' cities
-#'
-#' # Unest to get one row per related page:
-#' tidyr::unnest(cities, "related")
+#' get_page_langlinks("Charles Harpur", failure_mode = "quiet")
 #'
 #' # The functions are vectorised over title and language
 #' # Find all articles about Joanna Baillie, and retrieve summary data for
 #' # the first two.
 #' baillie <- get_page_langlinks("Joanna Baillie") %>%
 #'   dplyr::slice(1:2) %>%
-#'   dplyr::mutate(get_page_summary(title = title, language = code))
+#'   dplyr::mutate(get_page_summary(title = title, language = code, failure_mode = "quiet"))
 #' baillie
 NULL
 
 #' @rdname page_vector_functions
 #' @export
-get_latest_revision <- function(title, language = "en") {
+get_latest_revision <- function(title, language = "en", failure_mode = "error") {
   get_rest_resource(
     "page", "title", title,
-    language = language, api = "wikimedia", response_type = "revision_metadata"
+    language = language, api = "wikimedia", response_type = "revision_metadata",
+    failure_mode = failure_mode
   )
 }
 
@@ -75,20 +64,22 @@ parse_response.revision_metadata <- function(response) {
 
 #' @rdname page_vector_functions
 #' @export
-get_page_html <- function(title, language = "en") {
+get_page_html <- function(title, language = "en", failure_mode = "error") {
   get_rest_resource(
     "page", "html", title,
-    language = language, api = "wikimedia", response_format = "html"
+    language = language, api = "wikimedia", response_format = "html",
+    failure_mode = failure_mode
   )
 }
 
 #' @rdname page_vector_functions
 #' @export
-get_page_summary <- function(title, language = "en") {
+get_page_summary <- function(title, language = "en", failure_mode = "error") {
   get_rest_resource(
     "page", "summary", title,
     language = language, api = "wikimedia",
-    response_type = "summary"
+    response_type = "summary",
+    failure_mode = failure_mode
   )
 }
 
@@ -99,22 +90,7 @@ parse_response.summary <- function(response) {
 
 #' @rdname page_vector_functions
 #' @export
-get_page_related <- function(title, language = "en") {
-  get_rest_resource(
-    "page", "related", title,
-    language = language, api = "wikimedia",
-    response_type = "summary_array"
-  )
-}
-
-#' @export
-parse_response.summary_array <- function(response) {
-  purrr::map(response, "pages") %>% purrr::map(flatten_bind)
-}
-
-#' @rdname page_vector_functions
-#' @export
-get_page_talk <- function(title, language = "en") {
+get_page_talk <- function(title, language = "en", failure_mode = "error") {
   talk_pattern <- "^\\w+:"
   if (!all(stringr::str_detect(title, talk_pattern))) {
     rlang::abort("One or more titles do not begin with 'Talk:' or similar",
@@ -122,16 +98,18 @@ get_page_talk <- function(title, language = "en") {
   }
   get_rest_resource(
     "page", "talk", title,
-    language = language, api = "wikimedia"
+    language = language, api = "wikimedia",
+    failure_mode = failure_mode
   )
 }
 
 #' @rdname page_vector_functions
 #' @export
-get_page_langlinks <- function(title, language = "en") {
+get_page_langlinks <- function(title, language = "en", failure_mode = "error") {
   get_rest_resource(
     "page", title, "links", "language",
-    language = language, response_type = "page_language_object"
+    language = language, response_type = "page_language_object",
+    failure_mode = failure_mode
   )
 }